+++ automate-mautic.php
@@ -16,7 +16,7 @@
 if ( ! defined( 'ABSPATH' ) ) {
 	exit();
 }
-require_once 'classes/class-amp-mautic-loader.php';
+require_once 'classes/class-amp-loader.php';
 /*function ap_check_user_access() {
 	if( current_user_can( 'access_automate_mautic' ) ) {
 	}
